MAJOR CHANGE: Now I've got your attention... :P
I made a few changes to wrinklyninja-user-rules that removed the second parsing of the masterlist, and now BOSS runs roughly 10x faster than it did, at least for me. :D
I'd like some testing of this though, I have 225 mods and it used to take 9-10 seconds, but now it takes ~1 second. If it's not just me, then I'll merge the relevant changes into trunk.
I'll open up a code review for this.

git-svn-id: http://better-oblivion-sorting-software.googlecode.com/svn@1370 bfee223e-1fc4-11de-b650-e715bd6d7cf1