Added conditional-statements-in-masterlist support. Currently only checks whether mods exist or not. There are two syntaxes:
IF: <MOD> THEN: <MOD/MESSAGE>

Which will parse the <MOD/MESSAGE> if the <MOD> exists, acting as if the line only contained <MOD/MESSAGE>. If <MOD> doesn't exist, then the line will be ignored. There is also the opposite:

IFNOT: <MOD> THEN: <MOD/MESSAGE>

Which will ignore the line if <MOD> exists, and parse the line if it doesn't.

This is intended as a replacement to the current > and < symbols used. Implementation is a bit hacky since this sort of thing is best done through an actual parser rather than the current system. I'm also sure there's a more efficient way to do it.

git-svn-id: http://better-oblivion-sorting-software.googlecode.com/svn@2272 bfee223e-1fc4-11de-b650-e715bd6d7cf1