Code Reshuffle Part 1: Delete boss-common.
This is part of ongoing work to reorganise BOSS's code.
All common functionality will be stored in a new boss-common tree which compiles a static library.
CLI-specific code required to build a CLI version of the BOSS program will be stored in a boss-cli tree.
The boss-gui tree will store GUI-specific code for building a GUI version of the BOSS program.
The code for the BOSS API .dll will be stored in boss-api.

git-svn-id: http://better-oblivion-sorting-software.googlecode.com/svn@3244 bfee223e-1fc4-11de-b650-e715bd6d7cf1