This update to the BOSS-Common library breaks BOSS GUI compiling. Changes:
- Rewrote large swathes of BOSS to use new classes and their member functions for operations.
- Rewrote how user rules are applied that seems to have resulted in a small performance boost.
- Parsing grammars turned into classes and associated functions made private members.
- Combined Skipper and Ini_Skipper parsers.
- Grammar error reporting now uses new class to distance the content of error messages from the formatting.
- Removed some unused global messages.
- Most (if not all) looping flow controls now use iterators instead of element indexes.
- Some functions that used to return values to signify errors now use the boss_error class as is appropriate.
- Lots more...

Execute.h/cpp is still a bit of a mess, and the functions need to be re-evaluated to ensure they are what is really required. An update to Output.h/cpp is incoming, bringing a new class structure with overloaded operators, etc. for simpler outputting with formatting separate from content. The parsing evaluation functions in Grammar.h need to be updated to use the new classes and their member functions, after which the majority of Helpers.h can be discarded. Also need to enforce the trimming of class strings/paths so that Tidy() only needs to perform a to_lower conversion.

BOSS GUI will be updated to conform to the changes once the above work has been carried out.

git-svn-id: http://better-oblivion-sorting-software.googlecode.com/svn@3408 bfee223e-1fc4-11de-b650-e715bd6d7cf1