- Updated Russian translation. I don't understand it, but it looks cool. :P - Updated batch files so they should now stay open so you can see if anything went wrong. - Re-did GUI translation code, apparently Windows' UI doesn't use UTF-8 so was displaying gibberish when > ASCII code points were given to it.
Noticed that the GUI's User Rules Manager is now displaying the Masterlist tree list as being empty. Will have to back-track to see when that happened. I didn't notice any other GUI bugs.

CLI is still displaying gibberish when > ASCII code points are passed to it, unfortunately there is no way around this programmatically (I'm sure that's a word) that can be done using a cross-platform approach that doesn't use wxWidgets (it may be possible using that). I don't want to use wxWidgets for the CLI, because I equate wxWidgets in my mind with GUIs. It may be silly, but until I am otherwise educated, that's the way it is. As such, I'll just add this as a "known limitation" to the readme.

git-svn-id: http://better-oblivion-sorting-software.googlecode.com/svn@4796 bfee223e-1fc4-11de-b650-e715bd6d7cf1